fix: fixed Signing test
